Changes after running spotless apply mvn:apply

+++ src/main/java/payroll/EmployeeController.java
@@ -4,7 +4,6 @@ import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.*;
 
 import java.util.List;
 import java.util.stream.Collectors;
-
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.IanaLinkRelations;
@@ -23,23 +22,24 @@ class EmployeeController {
 
   private final EmployeeRepository repository;
   private final EmployeeModelAssembler assembler;
-  EmployeeController(EmployeeRepository repository,
-		  EmployeeModelAssembler assembler) {
+
+  EmployeeController(EmployeeRepository repository, EmployeeModelAssembler assembler) {
     this.repository = repository;
     this.assembler = assembler;
   }
-
 
   // Aggregate root
   // tag::get-aggregate-root[]
   @GetMapping("/employees")
   CollectionModel<EntityModel<Employee>> all() {
 
-    List<EntityModel<Employee>> employees = repository.findAll().stream() //
-        .map(assembler::toModel) //
-        .collect(Collectors.toList());
+    List<EntityModel<Employee>> employees =
+        repository.findAll().stream() //
+            .map(assembler::toModel) //
+            .collect(Collectors.toList());
 
-    return CollectionModel.of(employees, linkTo(methodOn(EmployeeController.class).all()).withSelfRel());
+    return CollectionModel.of(
+        employees, linkTo(methodOn(EmployeeController.class).all()).withSelfRel());
   } // end::get-aggregate-root[]
 
   @PostMapping("/employees")
@@ -52,30 +52,38 @@ class EmployeeController {
         .body(entityModel);
   }
   // Single item
-  
+
   @GetMapping("/employees/{id}")
   EntityModel<Employee> one(@PathVariable Long id) {
 
-     Employee employee = repository.findById(id) //
-         .orElseThrow(() -> new EmployeeNotFoundException(id));
+    Employee employee =
+        repository
+            .findById(id) //
+            .orElseThrow(() -> new EmployeeNotFoundException(id));
 
-     return assembler.toModel(employee);
+    return assembler.toModel(employee);
   }
-  
+
   @PutMapping("/employees/{id}")
   ResponseEntity<?> replaceEmployee(@RequestBody Employee newEmployee, @PathVariable Long id) {
 
-    Employee updatedEmployee = repository.findById(id) //
-        .map(employee -> {
-          if(newEmployee.getFirstName() != null )  employee.setFirstName(newEmployee.getFirstName());
-          if(newEmployee.getLastName() != null) employee.setLastName(newEmployee.getLastName());
-          if(newEmployee.getRole() != null) employee.setRole(newEmployee.getRole());
-          return repository.save(employee);
-        }) //
-        .orElseGet(() -> {
-          newEmployee.setId(id);
-          return repository.save(newEmployee);
-        });
+    Employee updatedEmployee =
+        repository
+            .findById(id) //
+            .map(
+                employee -> {
+                  if (newEmployee.getFirstName() != null)
+                    employee.setFirstName(newEmployee.getFirstName());
+                  if (newEmployee.getLastName() != null)
+                    employee.setLastName(newEmployee.getLastName());
+                  if (newEmployee.getRole() != null) employee.setRole(newEmployee.getRole());
+                  return repository.save(employee);
+                }) //
+            .orElseGet(
+                () -> {
+                  newEmployee.setId(id);
+                  return repository.save(newEmployee);
+                });
 
     EntityModel<Employee> entityModel = assembler.toModel(updatedEmployee);
 
@@ -83,12 +91,11 @@ class EmployeeController {
         .created(entityModel.getRequiredLink(IanaLinkRelations.SELF).toUri()) //
         .body(entityModel);
   }
-  
+
   @DeleteMapping("/employees/{id}")
   ResponseEntity<?> deleteEmployee(@PathVariable Long id) {
-	if(repository.existsById(id))
-        repository.deleteById(id);
-	 else throw new EmployeeNotFoundException(id);
-    return new ResponseEntity<>("Employee with id: "+id+" was deleted", HttpStatus.OK);
+    if (repository.existsById(id)) repository.deleteById(id);
+    else throw new EmployeeNotFoundException(id);
+    return new ResponseEntity<>("Employee with id: " + id + " was deleted", HttpStatus.OK);
   }
 }
